Changed key_len size_t type to uint8 so as to replicate bug for 2a prefix

+++ c_src/bcrypt_nif.c
@@ -54,7 +54,7 @@ static ERL_NIF_TERM bf_init(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[])
 	ErlNifBinary state;
 	char key[1024];
 	char salt[1024];
-	size_t key_len;
+	uint8_t key_len;
 	unsigned long key_len_arg;
 	uint8_t salt_len;
 
@@ -62,7 +62,7 @@ static ERL_NIF_TERM bf_init(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[])
 			!enif_get_ulong(env, argv[1], &key_len_arg) ||
 			!enif_get_string(env, argv[2], salt, sizeof(salt), ERL_NIF_LATIN1))
 		return enif_make_badarg(env);
-	key_len = (size_t) key_len_arg;
+	key_len = (uint8_t) key_len_arg;
 	salt_len = BCRYPT_MAXSALT;
 
 	if (!enif_alloc_binary(sizeof(blf_ctx), &state))
@@ -80,7 +80,7 @@ static ERL_NIF_TERM bf_expand(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[
 	ErlNifBinary state;
 	char key[1024];
 	char salt[1024];
-	size_t key_len;
+	uint8_t key_len;
 	unsigned long key_len_arg;
 	uint8_t salt_len;
 
@@ -90,7 +90,7 @@ static ERL_NIF_TERM bf_expand(ErlNifEnv* env, int argc, const ERL_NIF_TERM argv[
 			!enif_get_ulong(env, argv[2], &key_len_arg) ||
 			!enif_get_string(env, argv[3], salt, sizeof(salt), ERL_NIF_LATIN1))
 		return enif_make_badarg(env);
-	key_len = (size_t) key_len_arg;
+	key_len = (uint8_t) key_len_arg;
 	salt_len = BCRYPT_MAXSALT;
 
 	Blowfish_expand0state((blf_ctx *) state.data, (uint8_t *) key, key_len);
